fix: move non-action exports out of "use server" actions.ts so the app builds
`app/actions.ts` carries `"use server"`, where Next/Turbopack allows only
async-function exports. It also exported several plain values —
`REFRESH_INTERVAL_OPTIONS`, `isAllowedRefreshInterval`, `TAB_GROUP_MAX`,
`COLOR_HEX_RE`, `normalizeColumnColor`, `DECK_EXPORT_VERSION` — which trips
"Only async functions are allowed to be exported in a 'use server' file".
That makes the whole module fail to load, cascading into 62 build errors
("The module has no exports at all") across every importer, so `next build`
fails outright. `tsc` does not enforce this rule, which is how it landed on a
green-looking main.

Hoist the shared constants + sync validators into a new plain module
`lib/deck-rules.ts`, import them back into `actions.ts` for internal use, and
point the client store at the new module. `actions.ts` now exports only Server
Actions (plus erased type-only exports, which "use server" permits).

Verified: `next build` compiles + generates all routes, `tsc --noEmit` clean,
eslint clean.
